chore: send all accounts with their tokens
