+++ packages/wallets/src/components/WalletsAddMoreCardBanner/WalletsAddMoreCardBanner.tsx
@@ -1,15 +1,10 @@
 import React, { useEffect } from 'react';
-import Cookies from 'js-cookie';
 import { useHistory } from 'react-router-dom';
-import { useActiveWalletAccount, useCreateWallet, useIsEuRegion } from '@deriv/api-v2';
+import { useCreateWallet, useIsEuRegion } from '@deriv/api-v2';
 import { LabelPairedCheckMdFillIcon, LabelPairedPlusMdFillIcon } from '@deriv/quill-icons';
-import { getAccountsFromLocalStorage } from '@deriv/utils';
-import { Analytics } from '@deriv-com/analytics';
 import { Localize, useTranslations } from '@deriv-com/translations';
 import { Button, useDevice } from '@deriv-com/ui';
-import { URLConstants } from '@deriv-com/utils';
-import { LANDING_COMPANIES } from '../../constants/constants';
-import { isProduction, OUT_SYSTEMS_TRADERSHUB } from '../../helpers/urls';
+import { redirectToOutSystems } from '../../helpers/urls';
 import useSyncLocalStorageClientAccounts from '../../hooks/useSyncLocalStorageClientAccounts';
 import useWalletAccountSwitcher from '../../hooks/useWalletAccountSwitcher';
 import { TWalletCarouselItem } from '../../types';
@@ -32,7 +27,6 @@ const WalletsAddMoreCardBanner: React.FC<TWalletCarouselItem> = ({
     const { addWalletAccountToLocalStorage } = useSyncLocalStorageClientAccounts();
     const { localize } = useTranslations();
     const { data: isEuRegion } = useIsEuRegion();
-    const { data: activeWallet } = useActiveWalletAccount();
 
     useEffect(
         () => {
@@ -69,35 +63,6 @@ const WalletsAddMoreCardBanner: React.FC<TWalletCarouselItem> = ({
         ]
     );
 
-    const redirectToOutSystems = () => {
-        // redirect to OS Tradershub if feature is enabled
-        const isOutSystemsRealAccountCreationEnabled = Analytics?.getFeatureValue(
-            'trigger_os_real_account_creation',
-            false
-        );
-        if (isOutSystemsRealAccountCreationEnabled) {
-            const clientAccounts = getAccountsFromLocalStorage() ?? {};
-            const loginid = activeWallet?.loginid ?? '';
-            if (!loginid) return;
-            const authToken = clientAccounts[loginid].token;
-            if (!authToken) return;
-            const expires = new Date(new Date().getTime() + 1 * 60 * 1000); // 1 minute
-
-            Cookies.set('os_auth_token', authToken, { domain: URLConstants.baseDomain, expires });
-
-            const params = new URLSearchParams({
-                action: 'real-account-signup',
-                currency: currency ?? '',
-                target: LANDING_COMPANIES.MALTAINVEST,
-            });
-            const baseUrl = isProduction() ? OUT_SYSTEMS_TRADERSHUB.PRODUCTION : OUT_SYSTEMS_TRADERSHUB.STAGING;
-
-            const redirectURL = new URL(`${baseUrl}/redirect`);
-            redirectURL.search = params.toString();
-            return (window.location.href = redirectURL.toString());
-        }
-    };
-
     return (
         <div className='wallets-add-more__banner'>
             <div className='wallets-add-more__banner-header'>
@@ -120,7 +85,7 @@ const WalletsAddMoreCardBanner: React.FC<TWalletCarouselItem> = ({
                     if (!currency) return;
 
                     if (isEuRegion) {
-                        return redirectToOutSystems();
+                        return redirectToOutSystems(currency);
                     }
 
                     const createAccountResponse = await mutateAsync({
