+++ packages/wallets/src/constants/index.ts
@@ -1,0 +1,4 @@
+export * from './constants';
+export * from './errorCodes';
+export * from './icons';
+export * from './password';
