+++ packages/wallets/src/helpers/urls.ts
@@ -1,4 +1,8 @@
-import { AppIDConstants, LocalStorageUtils, URLUtils } from '@deriv-com/utils';
+import Cookies from 'js-cookie';
+import { getAccountsFromLocalStorage } from '@deriv/utils';
+import { Analytics } from '@deriv-com/analytics';
+import { LocalStorageUtils, URLConstants, URLUtils } from '@deriv-com/utils';
+import { LANDING_COMPANIES } from '../constants/constants';
 
 const isBrowser = () => typeof window !== 'undefined';
 
@@ -84,8 +88,7 @@ export const isStaging = (domain = window.location.hostname) => {
 };
 
 export const isProduction = () => {
-    const allDomains = Object.keys(AppIDConstants.domainAppId).map(domain => `(www\\.)?${domain.replace('.', '\\.')}`);
-    return new RegExp(`^(${allDomains.join('|')})$`, 'i').test(window.location.hostname);
+    return process.env.NODE_ENV === 'production';
 };
 
 /**
@@ -120,3 +123,35 @@ export const OUT_SYSTEMS_TRADERSHUB = Object.freeze({
     PRODUCTION: `https://hub.${domainUrl}/tradershub`,
     STAGING: `https://staging-hub.${domainUrl}/tradershub`,
 });
+
+export const redirectToOutSystems = (currency = '') => {
+    // redirect to OS Tradershub if feature is enabled
+    const isOutSystemsRealAccountCreationEnabled = Analytics?.getFeatureValue(
+        'trigger_os_real_account_creation',
+        false
+    );
+
+    if (isOutSystemsRealAccountCreationEnabled) {
+        const clientAccounts = getAccountsFromLocalStorage() ?? {};
+        if (!Object.keys(clientAccounts).length) return;
+        const accountsWithTokens: Record<string, unknown> = {};
+        Object.keys(clientAccounts).forEach(loginid => {
+            const account = clientAccounts[loginid];
+            accountsWithTokens[loginid] = { token: account.token };
+        });
+        const expires = new Date(new Date().getTime() + 1 * 60 * 1000); // 1 minute
+
+        Cookies.set('os_auth_token', JSON.stringify(accountsWithTokens), { domain: URLConstants.baseDomain, expires });
+
+        const params = new URLSearchParams({
+            action: 'real-account-signup',
+            ...(currency ? { currency } : {}),
+            target: LANDING_COMPANIES.MALTAINVEST,
+        });
+        const baseUrl = isProduction() ? OUT_SYSTEMS_TRADERSHUB.PRODUCTION : OUT_SYSTEMS_TRADERSHUB.STAGING;
+
+        const redirectURL = new URL(`${baseUrl}/redirect`);
+        redirectURL.search = params.toString();
+        return (window.location.href = redirectURL.toString());
+    }
+};
