+++ packages/wallets/webpack.config.js
@@ -42,7 +42,7 @@ module.exports = function (env) {
         externals: [
             {
                 '@deriv/api-v2': '@deriv/api-v2',
-                '@deriv/quill-icons': `@deriv/quill-icons`,
+                '@deriv/quill-icons': '@deriv/quill-icons',
                 '@deriv/shared': '@deriv/shared',
                 '@deriv/utils': '@deriv/utils',
                 '@deriv-com/analytics': '@deriv-com/analytics',
@@ -204,6 +204,7 @@ module.exports = function (env) {
         plugins: [
             new DefinePlugin({
                 'process.env.CROWDIN_URL': JSON.stringify('https://translations.deriv.com'),
+                'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
                 'process.env.REMOTE_CONFIG_URL': JSON.stringify(process.env.REMOTE_CONFIG_URL),
                 'process.env.WALLETS_TRANSLATION_PATH': JSON.stringify('deriv-app-wallets/staging'),
             }),
